Hastag Service 수정

+++ src/main/kotlin/sogong/sogongSpring/service/HashtagService.kt
@@ -27,7 +27,7 @@ class HashtagService {
     @Transactional
     fun savePostHashtag(postHashtagDto: PostHashtagDto){
         val hashIds = hashtagDbRepository.findByHashNames(postHashtagDto.hashName)
-        if(hashIds.isNotEmpty()) {
+        if(hashIds.size == postHashtagDto.hashName.size) {
             runCatching{
                 entirePostRepository.findById(postHashtagDto.postId).get()
             }.onSuccess { epe ->
@@ -41,50 +41,54 @@ class HashtagService {
 
     @Transactional
     fun saveUserHashtag(userHashTagDto: UserHashTagDto){
-        runCatching{
-            userLoginRepository.findById(userHashTagDto.userId).get()
-        }.onSuccess {
-            if(hashtagDbRepository.findByHashNames(userHashTagDto.hashName).isEmpty())
-                throw HashNameException(userHashTagDto.hashName)
-            else{
-                for(hash in userHashTagDto.hashName)
-                    userHashtagRepository.save(
-                        UserHashtagEntity(userId = it, hashName = hash, groupId = userHashTagDto.groupId)
-                    )
-            }
-        }.onFailure { throw UserIdException(userHashTagDto.userId)}
+        if(hashtagDbRepository.findByHashNames(userHashTagDto.hashName).size == userHashTagDto.hashName.size) {
+            runCatching {
+                userLoginRepository.findById(userHashTagDto.userId).get()
+            }.onSuccess {
+                if (hashtagDbRepository.findByHashNames(userHashTagDto.hashName).isEmpty())
+                    throw HashNameException(userHashTagDto.hashName)
+                else {
+                    for (hash in userHashTagDto.hashName)
+                        userHashtagRepository.save(
+                            UserHashtagEntity(userId = it, hashName = hash)
+                        )
+                }
+            }.onFailure { throw UserIdException(userHashTagDto.userId) }
+        } else throw HashNameException(userHashTagDto.hashName)
     }
 
     @Transactional
     fun editUserHashtag(userHashTagDto: UserHashTagDto){
-        runCatching{
-            userLoginRepository.findById(userHashTagDto.userId).get()
-        }.onSuccess {
-            val uhrs = userHashtagRepository.findByUserIdAndGroupId(it, userHashTagDto.groupId)
-            val hash = userHashTagDto.hashName
-            var difAndSize = uhrs.size-hash.size
+        if(hashtagDbRepository.findByHashNames(userHashTagDto.hashName).size == userHashTagDto.hashName.size){
+            runCatching{
+                userLoginRepository.findById(userHashTagDto.userId).get()
+            }.onSuccess {
+                val uhrs = userHashtagRepository.findByUserId(it)
+                val hash = userHashTagDto.hashName
+                var difAndSize = uhrs.size-hash.size
 
-            if(difAndSize>0){ //ex)해시태그 4->2개로
-                for(i:Int in 1..difAndSize)
-                    userHashtagRepository.delete(uhrs[uhrs.size-i])
-                difAndSize = hash.size-1
-            }
-            else if (difAndSize<0){ //ex)해시태그 2->5개로
-                for(i:Int in 1..abs(difAndSize)){
-                    userHashtagRepository.save(
-                        UserHashtagEntity(
-                            userId = it, groupId = userHashTagDto.groupId,
-                            hashName = userHashTagDto.hashName[hash.size-i]))
+                if(difAndSize>0){ //ex)해시태그 4->2개로
+                    for(i:Int in 1..difAndSize)
+                        userHashtagRepository.delete(uhrs[uhrs.size-i])
+                    difAndSize = hash.size-1
                 }
-                difAndSize = uhrs.size-1
-            }
-            else difAndSize=uhrs.size-1 //ex)해시태그 2->2개로
+                else if (difAndSize<0){ //ex)해시태그 2->5개로
+                    for(i:Int in 1..abs(difAndSize)){
+                        userHashtagRepository.save(
+                            UserHashtagEntity(
+                                userId = it,
+                                hashName = userHashTagDto.hashName[hash.size-i]))
+                    }
+                    difAndSize = uhrs.size-1
+                }
+                else difAndSize=uhrs.size-1 //ex)해시태그 2->2개로
 
-            for (i:Int in 0..difAndSize){
-                uhrs.get(i).hashName = hash.get(i)
-                userHashtagRepository.save(uhrs.get(i))
-            }
-        }.onFailure {throw UserIdException(userHashTagDto.userId)}
+                for (i:Int in 0..difAndSize){
+                    uhrs.get(i).hashName = hash.get(i)
+                    userHashtagRepository.save(uhrs.get(i))
+                }
+            }.onFailure {throw UserIdException(userHashTagDto.userId)}
+        } else throw HashNameException(userHashTagDto.hashName)
     }
 
     @Transactional
@@ -92,7 +96,7 @@ class HashtagService {
         val original = postHashtagRepository.findAllByPost(postHashtagDto.postId)
         if(original.isNotEmpty()) {
             val hashes = hashtagDbRepository.findByHashNames(postHashtagDto.hashName)
-            if(hashes.isNotEmpty()) {
+            if(hashes.size == postHashtagDto.hashName.size) {
                 for (i: Int in 0 until original.size)
                     postHashtagRepository.save(
                         PostHashtagEntity(
@@ -106,11 +110,25 @@ class HashtagService {
     }
 
     @Transactional
-    fun searchOrPost(hashtags: List<String>) : List<EntirePostEntity> {
+    fun searchBarPost(hashtags: List<String>) : List<EntirePostEntity> {
         val hashIds = hashtagDbRepository.findByHashNames(hashtags)
-        if(hashIds.isNotEmpty())
+        if(hashIds.size == hashtags.size)
             return postHashtagRepository.findByHashIds(hashIds) //Dto로 변경?
         else throw HashNameException(hashtags)
+    }
+
+    @Transactional
+    fun hashBoardPost(userId: Long) : Any{
+        runCatching{
+            userLoginRepository.findById(userId).get()
+        }.onSuccess { ulr ->
+            val uhr = userHashtagRepository.findByUserId(ulr)
+            if(uhr.isNotEmpty())
+                return searchBarPost(uhr.map { name -> name.hashName })
+            else
+                return emptyList<String>()
+        }.onFailure { return UserIdException(userId).message ?: 0 }
+        throw IllegalStateException("Server Error!!")
     }
 
     @Transactional
@@ -121,11 +139,11 @@ class HashtagService {
     }
 
     @Transactional
-    fun printUserHashtag(userId: Long, groupId: Long) : List<UserHashtagEntity>{
+    fun printUserHashtag(userId: Long) : List<UserHashtagEntity>{
         runCatching{
             userLoginRepository.findById(userId).get()
         }.onSuccess {
-            return userHashtagRepository.findByUserIdAndGroupId(it, groupId)
+            return userHashtagRepository.findByUserId(it)
         }.onFailure { throw UserIdException(userId)}
         throw IllegalStateException("Server Error!!")
     }
